Initial commit of functioning code
The code works without crashing for a 10 ms simulation. The core
components are included: included are modifications to runEulerCode,
DiscrAirgunShuttleMulti (capable of running both the original model and
the updated +shuttle+bubble model), and a shuttle dynamics function
shuttleEvolve. The configuration file is changed to accommodate the port
area.

Note: the wrapper for running the simulation is not included yet.

Warning: The "original model" setting does not reproduce the results of
[Watson 2019]. Probably an original setting or line of code was
overwritten. This will be fixed in a later commit.