Add experimental sectioned 1D flow model
This commit adds an experimental file that sections the 1D flow into
an area contraction section, a pressure drop, and then another area
contraction. This allows the modeling of multiple phenomena, e.g., mass
loss and energy loss, in addition to area contraction.

The issue with this model is the lack of state change correlations for
compressible flow, despite the availability of pressure drop relations for
incompressible flow (e.g., for 90-degree bends).