Add acoustic DDE toy model, update airgun model
A delay differential formulation of the acoustics equation coupled to the
shuttle dynamics equation is added. The correctness is to be verified. The
figs are using a partial solution, valid only after one reflection; the
code is changed to add sufficiently many images for correctness in a
finite domain of dependence.

The airgun model is updated to allow for subsonic and chamber-limited
cases in addition to the usual airgun-limited cases (which remains the
bulk of the simulation). A new "phase map" figure is produced by the
launch script, intended to illustrate which regime the flow was in during
the simulation. A plug flow model for the flow allowed by the displacement
of the shuttle is introduced. The assumption of flow into the plug flow at
speed equal to the shuttle velocity is problematic, since it causes
significant negative velocities for the chamber exit.

A coulomb damping with coefficient 0.3 is added for the solid-solid dry
friction. It looks negligible. Note that for a 3e4 linear damping
coefficient, e.g., arising from the constricted airflow in the operating
chamber acting as an air dashpot, a small number of oscillations occur in
the shuttle dynamics, constituting a plausible motion pattern.