Change piecewise RHS, correct DDE pressure reflection solution
The piecewise RHS (i.e., the models for the compressible flow divided into
cases based on where the flow potentially chokes) is modified. A mass flow
term into the shuttle "plug flow" section is added briefly and then
commented out since it caused negative mass flow rates when it wasn't
supposed to. The mass flow may be negligible; a flow term based on the
pressure difference between the airgun output p_a and the plug flow p_flow
may be considered next.

The difficulty with consistent choked flow states concerns the weak
enforcement of boundary conditions in the PDE domain. Since neither the
boundary, nor the second node in, respect the boundary condition
pointwise, we cannot use the SBP representation of the velocity field of
u_R to determine which flow state the port region is in. Possible
solutions are: (1) back out u_R using some functional acting on the last
few nodes of the SBP domain (based on SBP theory), or (2) keep track of
the boundary condition independently so that u_R is continuous in time.

See the phase diagram in this commit for an example trajectory. The kink
between the port-limited and chamber-limited flows is likely caused by the
weak enforcement issue.

This commit's subsonic code doesn't work yet: it seems the same issue is
preventing any meaningful results.

DDE:

The pressure reflection solution for a finite length tube requires summing
up all (positive) images of the Green's function in the domain of
dependence, rather than just one. The resulting simulation results are in
AcousticDDE/. The correct solution is implemented in this commit.