Add AcousticDDE resonance study, pressure-loss, ramp-up
Added resonance length study; resonance was found to have a bounded effect
on the pressure at the shuttle (firing chamber side). The resonance
features resemble the Duffing oscillator qualitatively.

Introduced exponential depressurization. At large depressurization
timescales, the damping is insufficient to bring the port to close. At
small timescales, the damping is too large and below a critical timescale,
the port is unable to open. (This critical timescale is investigated in
critical_timescale.m). At the appropriate timescale, the pressure is
sufficient to bring the shuttle to a close.

A proxy model for the pressure ``behind" the shuttle (i.e., in the
operating chamber but on the other side of the air cushion w.r.t. the
piston) is introduced to estimate the ``sticking" effect.